add extra link as part of AST render
