+++ code/go/site/static/css/site.css
@@ -4959,6 +4959,10 @@ html:has(.drawer-toggle:checked) {
   padding-bottom: 2rem;
 }
 
+.pl-1 {
+  padding-left: 0.25rem;
+}
+
 .pl-2 {
   padding-left: 0.5rem;
 }
